requirements.txt
Mako==1.0.0
pyramid==1.5.1
pyramid-debugtoolbar==2.1
pyramid-mako==1.0.2
python-bitcoinlib==0.1.1
waitress==0.8.9
